Create AccountSettings,PrivacySettings ,NotificationSettings and ThemeSelector
